Rename fetch_sep to fetchsep
