+++ fetchsep/json/templates/observations_example.json
