+++ fetchsep/json/templates/triggers_template.json
